feat(pages/index): atualizacao de busca das informacoes por id dos eventos - @jcdantss - @pablohunter
